move dropdown focus concerns from hook into CustomDropdown
